third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/inline/getAll.html

<!doctype html>
<meta charset="utf-8">
<title>StylePropertyMap.getAll tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#stylepropertymap">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<div id="target" style="width: 10px; --foo: auto; transition-duration: 1s, 2s;"></div>
<script>
'use strict';

const styleMap = document.getElementById('target').attributeStyleMap;

test(() => {
  assert_throws(new TypeError(), () => styleMap.getAll('lemon'));
}, 'Calling StylePropertyMap.getAll with an unsupported property throws a TypeError');

test(() => {
  const result = styleMap.getAll('height');
  assert_equals(result.length, 0);
}, 'Calling StylePropertyMap.getAll with a property not in the property model returns an empty list');

test(() => {
  const result = styleMap.getAll('--Foo');
  assert_equals(result.length, 0);
}, 'Calling StylePropertyMap.getAll with a custom property not in the property model returns an empty list');

test(() => {
  const result = styleMap.getAll('width');
  assert_style_value_array_equals(result, [CSS.px(10)]);
}, 'Calling StylePropertyMap.getAll with a valid property returns a single element list with the correct entry');

test(() => {
  const result = styleMap.getAll('wIdTh');
  assert_style_value_array_equals(result, [CSS.px(10)]);
}, 'StylePropertyMap.getAll is case-insensitive');

test(() => {
  const result = styleMap.getAll('--foo');
  assert_style_value_array_equals(result, [new CSSUnparsedValue('auto')]);
}, 'Calling StylePropertyMap.getAll with a valid custom property returns a single element list with the correct entry');

test(() => {
  const result = styleMap.getAll('transition-duration');
  assert_style_value_array_equals(result, [CSS.s(1), CSS.s(2)]);
}, 'Calling StylePropertyMap.getAll with a list-valued property returns all the values');

</script>
